refactor: move RunnerInput into a new module
